Display skeleton slides while available plans are being loaded
