+++ src/components/PlansSlider/index.js
@@ -1,0 +1,1 @@
+export { PlansSlider } from "./PlansSlider";
